moved lda to top-level LDA dir
